Include sources from Sfm-Toy-Library
As credited in the opening comments, I am using Roy Shilkrot's toy library as
an example implementation for producing depth maps using structure from motion.
He has a working implementation in his repository that I have moved here.

Syntactically it is correct, although the program does nothing at all at the
moment. The point was to move his code, which is meant as an example and is
thus in my opinion very convoluted, into this repository in order to test how
it works and, if I like it, refactor it to do only what I want - recover depth
from images. Now I will be working on it to simplify it and extract the
meaningful parts that I need.