Rework block matching tuner to be compatible with BlockMatcher classes
The ``StereoBMTuner`` is now a ``BMTuner`` that autodiscovers settable
parameters in the ``BlockMatcher`` that is passed to it.

This makes the code a little shorter and simpler. As long as the
``BlockMatcher`` the ``BMTuner`` receives fulfills the required API
requirements, the ``BMTuner`` can interact with it and creates the
appropriate toolbars needed to tune the ``BlockMatcher``.